[IMP] mail, portal: access params in the mail fetch route
This is a preparation PR to replace `/mail/chatter_fetch` with
`/mail/thread/messages`. This change assumes that the route can be accessed by
non internal users. This means that some access params could be sent to the
route, and we need to control the domain of the fetch.
